move to felix directory
