+++ felix/nfnetlink/nflog_linux.go
@@ -18,7 +18,6 @@
 package nfnetlink
 
 import (
-	"bytes"
 	"encoding/binary"
 	"fmt"
 	"syscall"
@@ -27,11 +26,10 @@ import (
 	"github.com/prometheus/client_golang/prometheus"
 	log "github.com/sirupsen/logrus"
 	"github.com/vishvananda/netlink/nl"
-	"golang.org/x/sys/unix"
 
+	"github.com/projectcalico/calico/felix/nfnetlink/nfnl"
+	"github.com/projectcalico/calico/felix/nfnetlink/pkt"
 	"github.com/projectcalico/calico/libcalico-go/lib/logutils"
-	"github.com/projectcalico/calico/nfnetlink/nfnl"
-	"github.com/projectcalico/calico/nfnetlink/pkt"
 )
 
 const (
@@ -94,16 +92,6 @@ func init() {
 var (
 	rll = logutils.NewRateLimitedLogger(logutils.OptInterval(time.Minute))
 )
-
-func SubscribeDNS(groupNum int, bufSize int, callback func(data []byte, timestamp uint64), done <-chan struct{}) error {
-	log.Infof("Subscribe to NFLOG group %v for DNS responses", groupNum)
-	resChan, err := openAndReadNFNLSocket(groupNum, bufSize, done, 2000, true, false)
-	if err != nil {
-		return err
-	}
-	parseAndReturnDNSResponses(groupNum, resChan, callback)
-	return nil
-}
 
 func NflogSubscribe(groupNum int, bufSize int, ch chan<- map[NflogPacketTuple]*NflogPacketAggregate, done <-chan struct{}, includeConnTrack bool) error {
 	resChan, err := openAndReadNFNLSocket(groupNum, bufSize, done, 2*cap(ch), false, includeConnTrack)
@@ -363,62 +351,6 @@ func parseAndAggregateFlowLogs(groupNum int, resChan <-chan [][]byte, ch chan<- 
 			}
 		}
 	}()
-}
-
-func parseAndReturnDNSResponses(groupNum int, resChan <-chan [][]byte, callback func(data []byte, timestamp uint64)) {
-	// Start another goroutine for parsing netlink messages into DNS response data.
-	go func() {
-		logCtx := log.WithFields(log.Fields{
-			"groupNum": groupNum,
-		})
-		logCtx.Debug("Start DNS response capture loop")
-		for {
-			select {
-			case res := <-resChan:
-				logCtx.Debugf("%v messages from DNS response channel", len(res))
-				for _, m := range res {
-					msg := nfnl.DeserializeNfGenMsg(m)
-					packetData, timestamp, err := getNflogPacketData(m[msg.Len():])
-					if err != nil {
-						logCtx.Warnf("Error parsing NFLOG %v", err)
-						continue
-					}
-					logCtx.Debugf("DNS response length %v", len(packetData))
-					callback(packetData, timestamp)
-				}
-			}
-		}
-	}()
-}
-
-func getNflogPacketData(m []byte) (packetData []byte, timestamp uint64, err error) {
-	var attrs [nfnl.NFULA_MAX]nfnl.NetlinkNetfilterAttr
-	n, err := nfnl.ParseNetfilterAttr(m, attrs[:])
-	if err != nil {
-		return
-	}
-	for idx := 0; idx < n; idx++ {
-		attr := attrs[idx]
-		attrType := int(attr.Attr.Type) & nfnl.NLA_TYPE_MASK
-		switch attrType {
-		case nfnl.NFULA_TIMESTAMP:
-			log.Debugf("DNS-LATENCY: NFULA_TIMESTAMP: %T %v", attr.Value, attr.Value)
-			var tv unix.Timeval
-			// NFLOG attributes are big-endian; see for example
-			// https://github.com/the-tcpdump-group/libpcap/blob/master/pcap/nflog.h
-			err := binary.Read(bytes.NewReader(attr.Value), binary.BigEndian, &tv)
-			if err != nil {
-				log.WithError(err).Panic("binary.Read failed")
-			}
-			log.Debugf("DNS-LATENCY: tv=%v", tv)
-			timestamp = uint64(tv.Usec*1000 + tv.Sec*1000000000)
-		case nfnl.NFULA_PAYLOAD:
-			packetData = attr.Value
-		default:
-			// Ignore attributes we don't care about.
-		}
-	}
-	return
 }
 
 func parseNflog(m []byte) (NflogPacket, error) {
